Add bluetooth Printing

+++ app/src/main/java/com/baqynra/withbaqyand/barbercut/Library.kt
@@ -4,7 +4,7 @@ import java.text.NumberFormat
 import java.util.*
 
 class Library {
-    fun toRupiah(nilai: Double): String? {
+    fun toRupiah(nilai: Int): String? {
         val localeID = Locale("in", "ID")
         val formatRupiah: NumberFormat = NumberFormat.getCurrencyInstance(localeID)
